+++ Changelog.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Comando ban
 - Comando kick
 - Comando bird
+- Comando pigeon
 - Task save_users que guarda la lista de usuarios
 - Script para construir y ejecutar en docker
 - Furry_irl en task memes